requirements.txt
streamlit
numpy
matplotlib
wordcloud
nltk
PyPDF2
Pillow
